foa_awdl: Split the runner and added an event queue.
