+++ foa_awdl/Cargo.toml
@@ -25,6 +25,7 @@ smoltcp = { version = "0.12.0", default-features = false, features = ["medium-et
 esp-config = { version = "0.3.1", features = ["build"] }
 
 [features]
+alloc = []
 esp32 = ["foa/esp32"]
 esp32s2 = ["foa/esp32s2"]
